+++ Tests/Fixtures/TestProject/build.sh
@@ -17,13 +17,6 @@ echo "
 xcodebuild -quiet -workspace Workspace.xcworkspace -scheme "App_iOS Test" -configuration "Test Debug" -xcconfig fixtures.xcconfig
 echo "✅ Successfully built iOS app"
 
-if [[ "$XCODE_VERSION" == 12* ]]; then
-    echo "
-    ⚙️ Building iOS app (Xcode 12+)"
-    xcodebuild -quiet -project ProjectXcode12.xcodeproj -scheme "App_iOS_With_Clip Test" -configuration "Test Debug" -xcconfig fixtures.xcconfig
-    echo "✅ Successfully built iOS app (Xcode 12+)"
-fi
-
 echo "
 ⚙️ Building macOS app"
 xcodebuild -quiet -workspace Workspace.xcworkspace -scheme "App_macOS" -configuration "Test Debug" -xcconfig fixtures.xcconfig