+++ scripts/gen-fixtures.sh
@@ -3,6 +3,5 @@ set -e
 
 swift run xcodegen --spec Tests/Fixtures/TestProject/AnotherProject/project.yml
 swift run xcodegen --spec Tests/Fixtures/TestProject/project.yml
-swift run xcodegen --spec Tests/Fixtures/TestProject/project-xcode-12.yml
 swift run xcodegen --spec Tests/Fixtures/CarthageProject/project.yml
 swift run xcodegen --spec Tests/Fixtures/SPM/project.yml